requirements.txt
absl-py==0.9.0
adal==1.2.2
astor==0.8.1
astroid==2.3.3
azure-cognitiveservices-search-imagesearch==2.0.0
azure-common==1.1.25
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageai==2.1.5
imagededup==0.2.2
imutils==0.5.3
isodate==0.6.0
isort==4.3.21
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
matplotlib==3.2.0
mccabe==0.6.1
msrest==0.6.11
msrestazure==0.6.2
numpy==1.16.6
oauthlib==3.1.0
object-detection==0.1
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
opt-einsum==3.2.0
pandas==1.0.2
Pillow==6.2.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pytesseract==0.3.3
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.0.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
tqdm==4.43.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
